Update README.md
Somewhat updated README file in association with issue #11. This is obviously just a starting point, with some placeholders. The contributor section should have every team member's name spelled right. Also notice that it is in alphabetical order for convenience. The uses is meant to be an indication of what the project is striving to be like, but this will later be updated with images and more instructions on how to use our tool when more progress on the project is made.