#22 Successful Approach
This commit features a successful approach to using the react app to send a prompt question to the FastAPI, which then queries the ChatGPT API. The response from ChatGPT API is then sent to the console log and is alerted to the user. Future commits should have the responses displayed on the website iteself, but this is good progress.