Improved documentation in certain areas
The API.py file had superfluous functions for the FastAPI that are not actually used. These were deleted, and documentation was added to the other code (with the exception of gemini code). In the main react app, documentation was added to the PromptForm.js component in order to improve readability. The only change made to the App.js file was the deletion of an empty space.