Adding ChakraUI to React components.
Restructured multiple components. Tried adding a transition UI feature when the light dark mode is toggled but could not figure it out yet. Overall pretty happy with what the current state is but, want to spend a bit more time on this. Also added new model buttons that don't do anything but they will be able to specify which AI model they want to test against.