#38 working functionality to parse files
The file parsing for pdf and docx works. If the user uploads a file, then the content is converted to text which is tested by sending that content to both chatgpt and gemini (producing queries). Some clean up and restructuring was also done to make the overall structure more clean. This also has the chakra code pulled from that branch mainly to test it visually.