#49 Some work done
A good chunk of work was done in terms of creating a model and automating the process of fine tuning it. The data set used is not a dataset relevant to our project, yet it was helpful in writing code that was independent of the dataset, which means that now we can use our own data set when it is ready since I generalized the code as much as possible. Of course, the portion of models.py where the model is being trained needs to be changed since it is very specific to the Australian rain csv file. This csv file will also be deleted before a pull request is made to main.