Moving all python related code into folder
All python related code is now in a subdirectory called python-code. This is something that will be cleaner, but it will also be easier to have all of these files in one area if we end up having time for making it into a package or docker image that can be run without needing the localhost.