Deleted old ml model files
These were files created while machine learning models were still being tested, but this is not accurate since they are not being used or even saved like this anymore (it also was wrong because they were not able to be reaccessed due to being saved with the wrong extension). The real ones are pickle files.