#76 Removed FileForm Component
We had a superfluous component that has already been integrated to the main form in terms of equivalent functionality. I went ahead and removed it and tested the app functionality that was already there to verify the changes did not break the app or cause any lost functionality. I also went ahead and changed the order of the buttons on the report page, mainly because I saw that on the main prompt page, the order was different and consistency is always good to keep.