#84 Removed the difficulty subdirectories
The leetcode problems have a difficulty ranking, but not all assignments that are non-leetcode will be judged on the same scale. Because of this, we cannot make a model that extracts this as a feature, and it needs to be more generic to allow for similar comparison of other files.