#83 This works now!
Now the user can submit feedback via the react app's thumbs up/down. This then adds the new datapoint to the dataset folder, which allows the model to be updated on this feedback in the future (since the model's dataset is created  every time a generate report request is made). At the same time, it is not feasible to have so many files if people spam it. Another reason why #80 would be good.