#83 Completed this
I refactored the code to make the structure easier to follow using a visual studio extension that formats js code (indentation). I also changed the way the documents were being added to the database because this was preventing the user feedback feature from working. I also made sure to implement the restriction of only labeling a piece of data once. This prevents users from spamming the thumbs up/down buttons in order to skew the dataset.