#72 Maximum prompt length accounted for
The maximum prompt length minimum for both api's is taken into account and a toast message is displayed if the maximum value is exceeded (without making any api call to prevent crashing). This error message is also time based, so it disappears and the user can try a different prompt after a short amount of time.