Now the report success notifies
Before the report generation button took the user to account home. This behavior has been updated so the user is notified the report generation could take several minutes, but they will see a notification on the bottom of their screen when their new report is ready. This is nicer behavior than how it was.